# api/ewbi/swagger.yaml
...
openapi: 3.0.3
info:
  version: 1.4.0
  title: Federation Management Service
  description: |
    # Introduction
    ---
    RESTful APIs that allow an OP to share the edge cloud resources and capabilities securely to other partner OPs over E/WBI.

    ---
    # API Scope

    ---
    APIs defined in this version of the specification can be categorized into the following areas:
    * __FederationManagement__ - Create and manage directed federation relationship with a partner OP
    * __AvailabilityZoneInfoSynchronization__ - Management of resources of partner OP zones and status updates
    * __ArtefactManagement__ - Upload, remove, retrieve and update application descriptors, charts and packages over E/WBI towards a partner OP
    * __FileManagement__ - Upload, remove, retrieve and update application binaries over E/WBI towards a partner OP
    * __ApplicationOnboardingManagement__ - Register, retrieve, update and remove applications over E/WBI towards a partner OP
    * __ApplicationDeploymentManagement__ - Create, update, retrieve and terminate application instances over E/WBI towards a partner OP
    * __AppProviderResourceManagement__ - Static resource reservation for an application provider over E/WBI for partner OP zones
    * __EdgeNodeSharing__ - Edge discovery procedures towards partner OP over E/WBI.
    * __LBORoamingAuthentication__ - Validation of user client authentication from home OP

    ---
    # Definitions
    ---
...

...
    * __OP__ - Operator platform.
...

...

...

...

...

...

...
components:
  schemas:
...
    Uri:
      type: string
...
    FederationRequestData:
      type: object
      required:
        - initialDate
        - partnerStatusLink
      properties:
        origOPFederationId:
          $ref: '#/components/schemas/FederationIdentifier'
        origOPCountryCode:
          $ref: '#/components/schemas/CountryCode'
        origOPMobileNetworkCodes:
          $ref: '#/components/schemas/MobileNetworkIds'
        origOPFixedNetworkCodes:
          $ref: '#/components/schemas/FixedNetworkIds'
        initialDate:
          type: string
          format: date-time
          description: Time zone info of the federation initiated by the originating OP
        partnerStatusLink:
          $ref: '#/components/schemas/Uri'
...
  responses:
...
    "401":
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    "409":
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    default:
      description: Generic Error
paths:
...
  /partner:
    post:
      summary: Creates one direction federation with partner operator platform.
      operationId: CreateFederation
      tags:
        - FederationManagement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FederationRequestData'
      responses:
        "200":
          description: Federation meta-info request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FederationResponseData'
          headers:
            Location:
              description: 'Contains the URI of the newly created resource, according to the structure: {apiRoot}/operatorplatform/federation/v1/partner/{federationContextId}'
              required: true
              schema:
                type: string
            Accept-Encoding:
              description: Accept-Encoding, described in IETF RFC 7694
              schema:
                type: string
            Content-Encoding:
              description: Content-Encoding, described in IETF RFC 7231
              schema:
                type: string
        "400":
          $ref: '#/components/responses/400BadRequest'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
      callbacks:
        onPartnerStatusEvent:
          '{$request.body#/partnerStatusLink }':
            post:
              requestBody:
                description: |
                  OP uses this callback api to notify partner OP about change in federation status, federation metadata or offered zone details. Allowed combinations of objectType and operationType are
                  - FEDERATION - STATUS: Status specified by parameter 'federationStatus'.
                  - ZONES - STATUS: Status specified by parameter 'zoneStatus'.
                  - ZONES - ADD: Use parameter 'addZones' to define add new zones
                  - ZONES - REMOVE: Use parameter 'removeZones' to define remove zones.
                  - EDGE_DISCOVERY_SERVICE - UPDATE: Use parameter 'edgeDiscoverySvcEndPoint' to specify new endpoints
                  - LCM_SERVICE - UPDATE: Use parameter 'lcmSvcEndPoint' to specify new endpoints
                  - MOBILE_NETWORK_CODES - ADD: Use parameter 'addMobileNetworkIds' to define new mobile network codes.
                  - MOBILE_NETWORK_CODES - REMOVE: Use parameter 'removeMobileNetworkIds' to remove mobile network codes.
                  - FIXED_NETWORK_CODES - ADD: Use parameter 'addFixedNetworkIds' to define new fixed network codes.
                  - FIXED_NETWORK_CODES - REMOVE: Use parameter 'removeFixedNetworkIds' to remove fixed network codes.
                  - SERVICE_APIS - ADD/REMOVE: Parameter Usage 'addServiceAPIs / removeServiceAPIs' to add or remove Service APIs support.
                content:
                  application/json:
                    schema:
                      type: object
                      required:
                        - federationContextId
                        - objectType
                        - operationType
                        - modificationDate
                      properties:
                        federationContextId:
                          $ref: '#/components/schemas/FederationContextId'
                        objectType:
                          type: string
                          enum:
                            - FEDERATION
                            - ZONES
                            - EDGE_DISCOVERY_SERVICE
                            - LCM_SERVICE
                            - MOBILE_NETWORK_CODES
                            - FIXED_NETWORK_CODES
                            - SERVICE_APIS
                        operationType:
                          type: string
                          enum:
                            - STATUS
                            - UPDATE
                            - ADD
                            - REMOVE
                        edgeDiscoverySvcEndPoint:
                          $ref: '#/components/schemas/ServiceEndpoint'
                        lcmSvcEndPoint:
                          $ref: '#/components/schemas/ServiceEndpoint'
                        addMobileNetworkIds:
                          $ref: '#/components/schemas/MobileNetworkIds'
                        removeMobileNetworkIds:
                          $ref: '#/components/schemas/MobileNetworkIds'
                        addFixedNetworkIds:
                          $ref: '#/components/schemas/FixedNetworkIds'
                        removeFixedNetworkIds:
                          $ref: '#/components/schemas/FixedNetworkIds'
                        addZones:
                          type: array
                          items:
                            $ref: '#/components/schemas/ZoneDetails'
                          description: List of zones, which the operator platform wishes to make available to developers/ISVs of requesting operator platform.
                          minItems: 1
                        removeZones:
                          type: array
                          items:
                            $ref: '#/components/schemas/ZoneIdentifier'
                          description: List of zones, which the operator platform no longer wishes to share.
                          minItems: 1
                        addServiceAPIs:
                          $ref: '#/components/schemas/serviceAPINames'
                        removeServiceAPIs:
                          $ref: '#/components/schemas/serviceAPINames'
                        zoneStatus:
                          type: array
                          items:
                            type: object
                            required:
                              - zoneId
                              - status
                            properties:
                              zoneId:
                                $ref: '#/components/schemas/ZoneIdentifier'
                              status:
                                $ref: '#/components/schemas/Status'
                          minItems: 1
                        federationStatus:
                          $ref: '#/components/schemas/Status'
                        modificationDate:
                          type: string
                          format: date-time
                          description: Date and time of the federation modification by the originating partner OP
              responses:
                "204":
                  description: Expected response to a successful call back processing
                "400":
                  $ref: '#/components/responses/400'
                "401":
                  $ref: '#/components/responses/401'
                "404":
                  $ref: '#/components/responses/404'
                "409":
                  $ref: '#/components/responses/409'
                "422":
                  $ref: '#/components/responses/422'
                "500":
                  $ref: '#/components/responses/500'
                "503":
                  $ref: '#/components/responses/503'
                "520":
                  $ref: '#/components/responses/520'
                default:
                  $ref: '#/components/responses/default'
              security:
                - notifClientCredentials: [fed-mgmt-notif]
...
  /{federationContextId}/artefact:
    post:
      summary: Uploads application artefact on partner OP. Artefact is a zip file containing scripts and/or packaging files like Terraform or Helm which are required to create an instance of an application.
      operationId: UploadArtefact
      tags:
        - ArtefactManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      requestBody:
        description: An application can consist of multiple components. App providers are allowed to define separate artefacts for each component or they could define a consolidated artefact at application level.
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - artefactId
                - appProviderId
                - artefactName
                - artefactVersionInfo
                - artefactVirtType
                - artefactDescriptorType
                - componentSpec
              properties:
                artefactNotifLink:
                  $ref: '#/components/schemas/Uri'
                artefactId:
                  $ref: '#/components/schemas/ArtefactId'
                appProviderId:
                  $ref: '#/components/schemas/AppProviderId'
                artefactName:
                  $ref: '#/components/schemas/ArtefactName'
                artefactVersionInfo:
                  $ref: '#/components/schemas/ArtefactVersionInfo'
                artefactDescription:
                  $ref: '#/components/schemas/ArtefactDescription'
                artefactVirtType:
                  $ref: '#/components/schemas/ArtefactVirtType'
                artefactFileName:
                  $ref: '#/components/schemas/ArtefactFileName'
                artefactFileFormat:
                  $ref: '#/components/schemas/ArtefactFileFormat'
                artefactDescriptorType:
                  $ref: '#/components/schemas/ArtefactDescriptorType'
                repoType:
                  $ref: '#/components/schemas/RepoType'
                artefactRepoLocation:
                  $ref: '#/components/schemas/ObjectRepoLocation'
                artefactFile:
                  type: string
                  format: binary
                  description: Helm archive/Terraform archive/container spec file or Binary image associated with an application component.
                componentSpec:
                  type: array
                  items:
                    $ref: '#/components/schemas/ComponentSpec'
                  minItems: 1
        required: true
      responses:
        "200":
          description: Artefact uploaded successfully
        "202":
          description: Artefact upload request accepted
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
...
  /{federationContextId}/files:
    post:
      summary: Uploads an image file. Originating OP uses this api to onboard an application image to partner OP.
      operationId: UploadFile
      tags:
        - ArtefactManagement
      parameters:
        - name: federationContextId
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/FederationContextId'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - fileId
                - appProviderId
                - fileName
                - fileVersionInfo
                - fileType
                - imgOSType
                - imgInsSetArch
              properties:
                fileNotifLink:
                  $ref: '#/components/schemas/Uri'
                fileId:
                  $ref: '#/components/schemas/FileId'
                appProviderId:
                  $ref: '#/components/schemas/AppProviderId'
                fileName:
                  $ref: '#/components/schemas/FileName'
                fileDescription:
                  $ref: '#/components/schemas/FileDescription'
                fileVersionInfo:
                  $ref: '#/components/schemas/FileVersionInfo'
                fileType:
                  $ref: '#/components/schemas/VirtImageType'
                checksum:
                  type: string
                  description: MD5 checksum for VM and file-based images, sha256 digest for containers
                imgOSType:
                  $ref: '#/components/schemas/OSType'
                imgInsSetArch:
                  $ref: '#/components/schemas/CPUArchType'
                repoType:
                  $ref: '#/components/schemas/RepoType'
                fileRepoLocation:
                  $ref: '#/components/schemas/ObjectRepoLocation'
                file:
                  type: string
                  format: binary
                  description: Binary image associated with an application component.
        required: true
      responses:
        "200":
          description: File uploaded successfully
        "202":
          description: File upload request accepted
        "400":
          $ref: '#/components/responses/400'
        "401":
          $ref: '#/components/responses/401'
        "404":
          $ref: '#/components/responses/404NotFound'
        "409":
          $ref: '#/components/responses/409'
        "422":
          $ref: '#/components/responses/422'
        "500":
          $ref: '#/components/responses/500'
        "503":
          $ref: '#/components/responses/503'
        "520":
          $ref: '#/components/responses/520'
        default:
          $ref: '#/components/responses/default'
...